+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPT6B8CS1TDGFS0Q4BR60J3E"
+        "01KPT9PFY2W1V4STBDHRR0QD1Q"
     )
     r"""The unique identifier of the data source"""