+++ packages/orq-rc/RELEASES.md
@@ -7219,3 +7219,13 @@ Based on:
 - [python v4.8.0-rc.45] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.45] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.45 - packages/orq-rc
+
+## 2026-04-22 09:56:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.46] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.46] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.46 - packages/orq-rc